Refactor subcommands to pass functions directly to constructor
This makes the class cleaner such that I don't have to declare redundant chatInputRun methods that just invoke existing functions.

Note that in order for this to work, the file for the base command has to be named after the command itself (e.g. the file for the base command cannot be index.ts). It will break otherwise.