Add script to pull changes on prod server when there are changes to the update script
The update script on Badstagram's server always has an extra unstaged change to my local version because it has a tag to give me permissions to run it. As such, I have to do this slightly annoying thing where I reset staged changes just to pull down current changes, then run this chmod command to add my permissions back. This change adds a script to do this for me.