Make overrides more specific
The lockfile always resolves to something like ^6.19.0, even when an override is specified. It resolves to whatever the dependencies field resolves to after running pnpm update or pnpm install, so it's best to ensure that the override also matches what is in the dependencies field, either specifying an exact version or one starting with ^, so that the --frozen-lockfile in actions works after updating.